你说得对，但是我的USB声卡没有反馈端点，感觉不如elipitic blep重采样

+++ HSDAC/User/usb/ch32v30x_usbhs_device.h
@@ -119,7 +119,7 @@ extern void MCU_Sleep_Wakeup_Operate (void);
 extern uint8_t USBHS_Endp_DataUp (uint8_t endp, uint8_t *pbuf, uint16_t len, uint8_t mod);
 
 extern uint32_t USBCDC_Write(const char* buf, uint32_t len);
-extern void USBUAC_WriteFeedback(uint32_t local_fs);
+extern void USBUAC_WriteFeedback(float local_fs);
 
 #ifdef __cplusplus
 }
